app/requirements.txt
asgiref==3.7.2
Django==5.0
django-redis==5.4.0
djangorestframework==3.14.0
Pillow==10.1.0
pytz==2023.3.post1
redis==5.0.1
sqlparse==0.4.4
